requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2

#for databases
SQLAlchemy==1.4.49
psycopg2

#for env var
python-dotenv
